+++ test/UT/StmtTest.hs
@@ -62,19 +62,43 @@ txnTest =
              ,(getTxnAsOf testTxns (toDate "20221216"))
              ,(getTxnAsOf testTxns (toDate "20221120"))
             ]
-      ,testCase "weight Average Balance ' " $ 
+      , let 
+          testTxns = [AccTxn (toDate "20221115") 100 20 Empty
+                      ,AccTxn (toDate "20221125") 90 (negate 10) Empty
+                      ,AccTxn (toDate "20221215") 80 (negate 10) Empty]
+        in 
+          testCase "Test View balance as of " $
+            assertEqual "View balance as of 1"
+            [80,100,100,80] $
+            [viewBalanceAsOf (toDate "20221114") testTxns,
+            viewBalanceAsOf (toDate "20221115") testTxns,
+            viewBalanceAsOf (toDate "20221116") testTxns,
+            viewBalanceAsOf (toDate "20221225") testTxns]
+
+      ,testCase "weight Average Balance 0 ' " $ 
+            assertEqual "Weight Average Balacne '"
+            0.27 $
+            weightAvgBalance' (toDate "20221115") (toDate "20221116") 
+              [BondTxn (toDate "20221115") 100 20 10 0.02 30 0 0 Nothing Empty ]
+      ,testCase "weight Average Balance 1" $ 
+            assertEqual "Weight Average Balacne '"
+            8.21 $
+            weightAvgBalance' (toDate "20221115") (toDate "20221215") 
+              [BondTxn (toDate "20221115") 100 20 10 0.02 30 0 0 Nothing Empty
+              ,BondTxn (toDate "20221215") 50 50 10 0.02 30 0 0 Nothing Empty]
+      ,testCase "weight Average Balance 2" $ 
             assertEqual "Weight Average Balacne '"
             14.74 $
             weightAvgBalance' (toDate "20221101") (toDate "20230101") 
               [BondTxn (toDate "20221115") 100 20 10 0.02 30 0 0 Nothing Empty
               ,BondTxn (toDate "20221215") 50 50 10 0.02 30 0 0 Nothing Empty]
-      ,testCase "weight Average Balance 2 ' " $ 
+      ,testCase "weight Average Balance 3" $ 
             assertEqual "Weight Average Balacne '"
             12.03 $
             weightAvgBalance' (toDate "20221110") (toDate "20230101")
               [(BondTxn (toDate "20221115") 100 20 10 0.02 30 0 0 Nothing Empty)
               ,(BondTxn (toDate "20221215") 50 50 10 0.02 30 0 0 Nothing Empty)]
-      ,testCase "weight Average Balance 3 ' " $ 
+      ,testCase "weight Average Balance 4" $ 
             assertEqual "Weight Average Balacne '"
             8.86 $
             weightAvgBalance'  (toDate "20220101") (toDate "20220201")