+++ ChangeLog.md
@@ -3,6 +3,9 @@
 ## 0.28.15
 ### 2024-07-15
 * NEW: pricing on asset with `rate curve`, pricing asset with `interest accrued`
+### 2024-7-31
+* FIX: enable compound formula on `weighted average` formula.
+
 
 ## 0.28.14
 ### 2024-07-06